more specs
